requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.21
django-appconf==1.1.0
django-ckeditor==6.7.2
django-compressor==4.5.1
django-debug-toolbar==5.2.0
django-extensions==4.1
django-imagekit==5.0.0
django-js-asset==2.2.0
django-meta==2.5.0
django-robots==6.1
django-seo==0.3.5
django-sitemaps==2.0.1
django-taggit==6.1.0
djangorestframework==3.16.0
kombu==5.5.3
lxml==5.4.0
pilkit==3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
rcssmin==1.1.2
redis==6.2.0
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
